Move language config to another file
